Tue, Aug 13, 2024, 5:59 PM +03:00

+++ filter-less/helpers.c
@@ -8,7 +8,7 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
     {
         for(int j = 0;j<width; j++)
         {
-            int avg_rgb=round(image[i][j].rgbtRed+image[i][j].rgbtGreen+image[i][j].rgbtBlue) / (float) 3;
+            int avg_rgb=round(image[i][j].rgbtRed+image[i][j].rgbtGreen+image[i][j].rgbtBlue) / 3.0;
             image[i][j].rgbtRed=image[i][j].rgbtGreen=image[i][j].rgbtBlue=avg_rgb;
 
         }
